+++ CHANGELOG.md
@@ -12,13 +12,15 @@
 
 ## v0.5.2
 
+- **New Feature**
+    + Add buttonAfter support (bootstrap skin) #126
 - **Documentation**
     + Add `attrs` option
     + Add Date input
     + Fix wrong imports in Customizations section
 - **Internal**
     + Optimise stuct and list re-rendering on value change
-    + server-side rendering markup differences #124
+    + fix server-side rendering markup differences #124
 
 ## v0.5.1